Fix module import error with python-docker symlink
Previously, virtualenv was installed using pip and it got the latest
release which currently as of Feb 13th 2020 is 20.0.3. This release has
a bug with module imports when you run a symlinked binary instead of the
full path.

As a solution, we're now apt installing virtualenv instead of installing
the latest release using pip.