Add become: true to task with become_user
The Ansible Galaxy recently flagged this as being a warning during
its linting check with ansible-lint and it degrades the quality of
this role.

This is a pretty ridiculous warning to have because most folks would set
become: true at the play level, not role level. I don't know of anyone
who explicitly sets become: true on every task instead of once at the
play level.

Maybe one day ansible-lint will evolve to understand that playbooks and
roles are used together.