Fix bug where empty configs would get skipped
Technically if you wanted to write out options that were empty
these tasks would get skipped instead of writing out an empty file.

Chances are you would want the empty options to be written out which is
basically removing whatever options you had.

Thanks to @coute for calling this out in:
  https://github.com/nickjj/ansible-docker/pull/115

That PR got stale (due to my fault) and needed quite a bit of work to
split it up so I'm committing it here directly with credit to the
author.